rimage: cast size_t to correct format types

+++ rimage/file_simple.c
@@ -375,8 +375,9 @@ static int simple_write_firmware(struct image *image)
 		return -errno;
 
 	fprintf(stdout, "firmware: image size %ld (0x%lx) bytes %d modules\n\n",
-			hdr.file_size + sizeof(hdr), hdr.file_size + sizeof(hdr),
-			hdr.num_modules);
+		(long) (hdr.file_size + sizeof(hdr)),
+		(long) (hdr.file_size + sizeof(hdr)),
+		hdr.num_modules);
 
 	return 0;
 }
@@ -459,7 +460,7 @@ int write_logs_dictionary(struct image *image)
 			fprintf(stdout, "logs dictionary: size %u\n",
 				header.data_length + header.data_offset);
 			fprintf(stdout, "including fw version of size: %lu\n\n",
-				sizeof(header.version));
+				(unsigned long) sizeof(header.version));
 		}
 	}
 out:
